Refactor test_many_to_many_descriptor
