+++ tests/end_to_end/test_many_to_many_descriptor.py
@@ -1,109 +1,6 @@
 from expects import expect, be_an, raise_error, have_property, equal, be_empty
 import json
-
-
-class ApiHelper:
-    @staticmethod
-    def post_a_skill(c, skill_text: str):
-        route = '/skills'
-        post_body = {
-            "text": skill_text
-        }
-        response = c.post(route, json=post_body)
-        body = json.loads(response.data)
-
-        expect(response.status_code).to(equal(201))
-        return body['id']
-
-    @staticmethod
-    def post_a_framework(c, skills: list):
-        route = '/frameworks'
-        post_body = {
-            "name": "test framework",
-            "skills": skills
-        }
-        response = c.post(route, json=post_body)
-        body = json.loads(response.data)
-
-        expect(response.status_code).to(equal(201))
-
-        return body['id']
-
-    @staticmethod
-    def get_a_framework_by_id(c, id: int):
-        route = '/frameworks'
-        response = c.get(f'{route}/{id}')
-        body = json.loads(response.data)
-
-        expect(response.status_code).to(equal(200))
-
-        return body
-
-    @staticmethod
-    def post_a_framework_with_no_skills(c):
-        route = '/frameworks'
-        post_body = {
-            "name": "test framework"
-        }
-        response = c.post(route, json=post_body)
-        body = json.loads(response.data)
-
-        expect(response.status_code).to(equal(201))
-
-        return body['id']
-
-    @staticmethod
-    def put_a_framework_skill(c, framework_id: int, skills: list):
-        route = f'/frameworks/{framework_id}/skills'
-        put_body = {
-            "skills": skills
-        }
-        response = c.put(route, json=put_body)
-        body = json.loads(response.data)
-
-        expect(response.status_code).to(equal(200))
-
-    @staticmethod
-    def patch_a_framework_skill(c, framework_id: int, skills_list: list):
-        route = f'/frameworks/{framework_id}/skills'
-        patch_body = {
-            "skills": skills_list
-        }
-        response = c.patch(route, json=patch_body)
-        body = json.loads(response.data)
-
-        expect(response.status_code).to(equal(200))
-
-    @staticmethod
-    def check_for_skills_on_framework(c, framework_id: int, skills_list: list):
-        route = f'/frameworks/{framework_id}/skills'
-        response = c.get(route)
-        body = json.loads(response.data)
-
-        expect(response.status_code).to(equal(200))
-        expect(body['skills']).to(equal(skills_list))
-
-    @staticmethod
-    def get_frameworks_on_skill(c, skill_id: int):
-        route = f'/skills/{skill_id}/frameworks'
-        response = c.get(route)
-        body = json.loads(response.data)
-
-        expect(response.status_code).to(equal(200))
-        return body
-
-    @staticmethod
-    def delete_a_framework_skill(c, framework_id: int, skills_list: list):
-        route = f'/frameworks/{framework_id}/skills'
-        delete_body = {
-            'skills': skills_list
-        }
-
-        response = c.delete(route, json=delete_body)
-        body = json.loads(response.data)
-
-        expect(response.status_code).to(equal(200))  # 204
-        return body
+from tests.service import ApiHelper
 
 
 class TestStartup():
